+++ packages/node/src/indexer/worker-fetch.module.ts
@@ -10,6 +10,7 @@ import {
   WorkerConnectionPoolStateManager,
   ConnectionPoolStateManager,
   NodeConfig,
+  InMemoryCacheService,
 } from '@subql/node-core';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { EthereumApiService } from '../ethereum';
@@ -20,6 +21,7 @@ import { IndexerManager } from './indexer.manager';
 import { ProjectService } from './project.service';
 import { SandboxService } from './sandbox.service';
 import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
+import { WorkerInMemoryCacheService } from './worker/worker.inMemoryCache.service';
 import { WorkerService } from './worker/worker.service';
 import { WorkerUnfinalizedBlocksService } from './worker/worker.unfinalizedBlocks.service';
 
@@ -72,6 +74,10 @@ import { WorkerUnfinalizedBlocksService } from './worker/worker.unfinalizedBlock
         new WorkerUnfinalizedBlocksService((global as any).host),
     },
     WorkerService,
+    {
+      provide: InMemoryCacheService,
+      useFactory: () => new WorkerInMemoryCacheService((global as any).host),
+    },
   ],
 })
 export class WorkerFetchModule {}